- Revised duration and attempts behaviour to initialize them when specified and update them depending on if its a new attempt or not. - Changed duration type from int to long.
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@13063 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ edu/xtec/adapter/Adapter.java
@@ -24,8 +24,8 @@ public interface Adapter
   public double getMaxGrade();
   public int getAttempts();
   public void setAttempts(int attempts);
-  public int getDuration();
-  public void setDuration(int duration);
+  public long getDuration();
+  public void setDuration(long duration);
   public String getLog();
   public void setLog(String log);
   public Vector getLogList();
